Add wait and sync steps to Gossip Server
Nodes part of the same Epoch are allowed
to talk to each other. When a new node enters
the cluster we want to prevent other nodes
from gossiping with it.

A new node will wait to receive what Era its
neighbours are in. Then wait for Era + 1 to join
in on the conversation