Don't proxy handle_info(nodedown, ...)
Instead of forcing the user to implement a handle_info call to receive
nodedown and nodeup messages we can check if a node has left the cluster
by comparing the list we have in state with the ErlangVM list. If a node
is missing we kow its been removed from the cluster and we can trigger a
down message on the callback module.